refactor: streamer page
